Moved things around and cleaning code.
