+++ Mito/App.js
@@ -1,12 +1,12 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
 import CheckboxClass from './classes/CheckboxClass.js';
-import { NativeBaseProvider, Link, Box } from "native-base"
+import HomeScreen from './classes/homeScreen.js';
+import { NativeBaseProvider} from "native-base"
 import {createNativeStackNavigator } from '@react-navigation/native-stack';
 import {NavigationContainer} from "@react-navigation/native";
 
 const MainStack = createNativeStackNavigator();
-export default function App(props) {
+
+export default function App() {
   return (
       <NavigationContainer>
     <NativeBaseProvider>
@@ -19,28 +19,4 @@ export default function App(props) {
   );
 }
 
-function HomeScreen(props){
-  return(
-  <View style={styles.container}>
-    <Text>Open up App.js to start working on your app!</Text>
-    <StatusBar style="auto" />
-    <Link onPress={() => props.navigation.navigate('Checkbox Class')} mt = '8'>
-      <Box px="3" py="2" bg="primary.400" rounded="sm" _text={{
-        color: "white",
-        fontWeight: "medium"
-      }}>
-        Press to go to your to do list
-      </Box>
-    </Link>
-  </View>
-  );
-}
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'left',
-  },
-});
